Fixed merge issues
A few issues popped up from the merge of Mark Branson's branch and the current master (~v1). The biggest issue revolved around the variable species_class, which was moved from the modal_aero_data module to the physpkg module. My first attempt to fix the issue resulted in a confusing circular dependency. I ended up moving the definition of species_class to the constituents module, although it is still given values in the physpkg module.