Fixed some SP run-time issues, problems persist
When using the 1-moment microphysics the "-chem none" option is the only one that currently works. The bulk aerosol option (trop_bam) causes a segmentation fault when the tracer data is being loaded and I'm not sure why. It probably has something to do with black carbon, and is probably related to the reasons that MAM3 is also not working.

When running the 2-moment microphysics in the CRM (SP2) the chemistry option can be set to "linoz_mam4_resus" or "linoz_mam4_resus_mom_soag", but in the latter case "-rain_evap_to_coarse_aero" must also be included in the configure arguments.

Although the trop_bam option is not working, it seems that the inclusion of namelist options for the linearized photo chemsitry (linoz) was causing problems. To make sure these options were excluded from the namelist I created a new "use case" option, 2000_cam5_av1c
Tho set this use case use:
    ./xmlchange --file env_run.xml --id CAM_NML_USE_CASE  --val 2000_cam5_av1c-SP1

Examples of valid configure arguments are:
    -use_SPCAM  -rad camrt -phys cam4 -nlev 26 -crm_nz 24 -crm_adv MPDATA -crm_nx 32 -crm_ny 1 -crm_dx 4000 -crm_dt 20  -SPCAM_microp_scheme sam1mom -cppdefs ' -DSP_DIR_NS ' -chem none
    -use_SPCAM  -rad rrtmg -phys cam5 -nlev 30 -crm_nz 28 -crm_adv MPDATA -crm_nx 32 -crm_ny 1 -crm_dx 4000 -crm_dt 20  -SPCAM_microp_scheme m2005   -cppdefs ' -DSP_DIR_NS ' -use_ECPP -chem linoz_mam4_resus_mom_soag -rain_evap_to_coarse_aero
Note that these have not been validated scientifically.

The "DSP_DIR_NS" option forces the 2D CRMs to be oriented meridionally.

The model was tested with higher vertical levels, but this resulted in an error,
    "ERROR: negative layer thickness.  timestep or remap time too large"