run-time issue - negative layer thickness
When trying to get ACME-SP to run with 72 levels I kept getting a floating point exception. I traced this down to the SurfaceResistanceMod in the land model. I was able to get past the error by restructuring an if statement block, but I'm very confused why this was happening for the SP model and not the standard model.

I'm also struggling to resolve a "negative layer thickness" error associated with the top 4 GCM levels that occurs in the first two timesteps.