Working with 72 layers, as lon ECPP is disabled
If ECPP is enabled the model will crash, usually with an "unreasonably high aerosol optical depth" error message. I also saw a lot of "negative layer thickness" errors, which might be related to a problem with the initial radiative heating profile. The negative thickness issue tends to be avoided when I use the SPRADBYPASS option, but both errors are reliably avoided by disabling the ECPP.

There is also a SPFLUXBYPASS option, which was provided by Mike Pritchard. This changes when surface fluxes are applied to the model state. without SPFLUXBYPASS the surface fluxes are added in the vertical diffusion solver, but these tendencies are not diffused when the CRM is active. The CRM still diffuses these fluxes, but without SPFLUXBYPASS the dynamical core is run in between when the fluxes are applied and when they are diffused. SPFLUXBYPASS changes this so that the fluxes are applied before the CRM call, so that the CRM can diffuse the fluxes before the dynamical core sees them.

Originally the SPFLUXBYPASS was tested with older SP-CAM using the bulk microphysics. With the 30-level ACME-SP this option appears to be crucial to run the model, otherwise a "unreasonably high aerosol optical depth" error appears near the surface. However, the current implementation bypass the surface flux of all constituents, but only adds the flux of water vapor to the CRM. This needs to be revisited so that we ensure the proper fluxes of all advected constituents.

There is also an ECPP_LEV_MOD option in this commit, which is meant to fix the issues with ECPP by limiting the calctulation to the levels of the CRM, but this is currently not working.