Adding routines to dump out data at beginning of crm(..)
On Monday, I'll add routines to dump out data at the end of crm(..). Currently, it's only for the first task, and I'll probably keep it that for now at least. Once I'm getting useful data, I may save more, but this is all that's needed to get the testbed up and running.

[BFB] - Bit-For-Bit