standalone executable compiling
Three main advances in this commit:

(1) setup_from_file.sh : You can now configure a build for standalone MMF by using one of the files dumped out during runtime, and this is the preferred way to do things. It's guaranteed to match the dimensions and configuration of the file being used

(2) crm_standalone.F90 : I now have an executable that runs the crm(...) call using data (although it isn't quite working yet), but it does build.

(3) Makefile : I updated the Makefile to include all the dependencies that aren't resolved through modules. Unfortunately there's a significant amount of code that doesn't live in modules, and the rouintes are just assumed to be resolved at link time. Not a great strategy, but it's what we've got.

I also updated the titan environment to use a newer cray-netcdf