Minor changes and clean-up
Nopthing major in this commit, but I did clean up some debugging output and added an error-trap bypass in the ZM code.
There is still code for the SPFLUXBYPASS options, but it is advised to not use this for now as it can cause issues that are not well understood.