Moved data dump routines to separate file
This commit includes several improvements. First, data dump routines are moved to a separate file so they aren't so distracting in crm_module.F90.

The data reading routines are changed so that they take a full "file.nc" rather than just a prefix, making it more convenient to code.

The setup.sh file is changes to use symbolic links so that changing code in the standalone directory changes the actual source, making development more convenient.

The data dumping routines in crm_dump.F90 are altered so that data is saved for only a CRM_DUMP_RATIO proportion of the crm(...) calls, and these are chosen randomly from a uniform random distribution. This reduces the total amount of data saved while maintaining representativeness of the saved data.