Miscelaneous improvements
Added ability to limit sampling size for columns to a smaller number via the -D_SAMP_LIMIT=[#] define.

Added --ovr to ncrcat command in collect_files.sh so that it doesn't ask you if you want to overwrite (because usually you do).

Added OpenACC flags to the compile commands for Scrollbuttons machine.