advect_scalar2D.F90 completely ported to GPUs; PCI-e transfers optimized
Confirmed the error norms over example columns are comparable to -O0 -O2 diffs for each output variable of crm(...).

There is a problem with efficiency, and it's likely either due to: (1) kernel serialization or (2) kernel launch latency

Of course, it could be both of these as well. I'm going to investigate what's going on with a profiler.