Adaptive surface flux smoothing
Even with the surface flux smoothing option, a recent run with crm_dx=4km crashed with a negative layer thickness error in the lowest model level over the Amazon. This is due to a 2 delta t oscillation in the surface fluxes. The cause of this numerical instability is uncertain, but it seems to be most strongly associated with rain forests. Since the surface flux smoothing option doesn't seem to be adequate when the crm grid spacing is large I decided to modify the smoother to adjust the relaxation timescale based on the residual fluxes. It has not been fully tested yet.