slight modifications
I increased the amount that my "adaptive" surface flux averaging modification changes the timescale, because a certain run keeps crashing.
There was also an issue with one of the Xhang-McFarlane error traps. I modified the error trap bypass to not have all the error messages filling up the log files.