Pushed loop over "ncols" down into the first level of CRM
I'm calling the number of columns inside the CRM "nvcols" for "number of vector columns", and the loop iteration index is "vc" for "vector column". Currently, I'm pushing all of "pcols" down into the CRM from the actual model and all of "ncols" down into the CRM for the standalone. Later, I'll put in some machinery to push an arbitrary-sized chunk.