CRM double precision harded coded and set by default
A problem in stratospheric temperature was identified and connected to the use of single precision real values in the CRM. Instead of addressing this with compiler flags it was decided to hardcode the option for double or single precision. Double precision is now the default in the CRM code, which is controlled by setting crm_rknd in params.F90. The preprocessor flag "CRM_SINGLE_PRECISION" overrides the default and reverts back to the original single precision. The problem can be reproduced by comparing single precision runs with crm timesteps of 10 and 1 sec. The differences between those single precision test simulations is almost completely eliminated in comparable runs with double precision.