Added capability for RRTMG with 1-moment micro
Following suggestions from Mark Branson, the call to m2005_effradius() needed to be replaced by cldefr() from pkg_cldoptics.F90 to enable RRTMG for the CRM 1-moment microphyscs. However, cldefr() operates on multiple columns, so to maintain consistency with how RRTMG operates with the 2-moment scheme I created a new subroutines cldefr_icol(), reitab_icol(), and reltab_icol() that operate on one column at a time. I think it might make more sense to change this in teh future so that both 1-mom and 2-mom configurations operate on a group of columns, but that seemed like a more invovled changes to the code, but I left a commented out block of code in radiation.F90 to this effect. I also had to zero out the cldfsnow and csnowp variables, although I'm not sure about the validity of this.

This modification also allows for using the conventional chemistry and aerosol options that are used in the non-SP model (i.e. -chem linoz_mam4_resus_mom_soag)

currently tests don't exists for this since it is a new configuration, so no formal testing procedure was done.