Fix declaration of get_reffc and get_reffi
Fix declaration of get_reffc and get_reffi functions in m2005
microphysics. The return type of these functions was declared twice,
once in the function declaration line and once in the function body,
causing the compiler to complain and die (Intel compiler on Skybridge).
Removing the type declaration from the function declaration line,
leaving just the declaration in the body (needed anyways to specify the
dimension of the return arrays) allows the 2-moment SP-CAM configuration
to build and run on skybridge. Fixes #5.