New SP component set names
The SP component set names were simplified. The new short names for the SP compsets are:
    FSP1V1
    FSP2V1
    FSP1V1-TEST         (reduced CRM domain for testing)
    FSP2V1-TEST
    F20TRSP1V1
    F20TRSP2V1

In the long compset definition "CAM%SP" is used instead of "CAM5%SP" to avoid the -phys flag, which is not needed. We don't foresore anyone wanting to revert to the old 1-moment configuration with 26 levels and CAMRT radiation (for several reasons), so we have not provided a compset for it.