Reduced Radiation calculation for RRTMG
This commit provides a modification to make SP simulations more economical by allowing the number of radiative calculations to be reduced.
Traditionally, radiation in SPCAM is calculated each time step of the GCM and for each time-averaged CRM column in order to allow for realistic cloud radiative feedbacks. The modification in this commit provides an alternative approach in which radiative tendencies are calculated from the average of a group of columns and the tendency is uniformally applied back to the group.
The variable crm_nx_rad (and also crm_ny_rad for 3D) controls how many radiative calculations are done, and must be a factor of crm_nx. if crm_nx_rad or crm_ny_rad is omitted then they are set to be equal to crm_nx and crm_ny, respectively, restoring the original behavior.
Some simple illustrative examples:
If crm_nx=64 and crm_nx_rad=2, then radiation will be calculated for 2 groups of 32 CRM columns, reducing the number of radiative calculations by a factor of 32.
If crm_nx=64 and crm_nx_rad=4, then radiation will be calculated for 4 groups of 16 CRM columns, reducing the number of radiative calculations by a factor of 16.