Replace "CAM" with "CAM5" to set options
Add the "5" back into the compset names (CAM%SP -> CAM5%SP) so that
options would be set properly. Replacing CAM5 with CAM seemed to set, in
particular, the CCSM_BGC variable to "none" as opposed to "CO2A", which
caused the SP simulations to crash. We might look into why this is, but
for the time being using "CAM5" allows our compsets to run without
modification.