Fix COSP height and pressure input
Fix COSP height and pressure inputs when copying CAM fields to COSP
input structure. Previously, the two second setup calls mixed up the
model level midpoints and interfaces, electing to use the model level
midpoints as the bottom interface and the model level interfaces tops as
the level midpoints for some strange reason. This was inconsistent with
the first setup call, and misleading. This changes bit-for-bit COSP
results.