Added eddy viscosity limiter in CRM
The limiter is contained in SGS_TKE/tke_full.F90 and enabled by SP_TK_LIM. The limiter has two parts. The first limits the buoyant production of TKE to be positive definite. The other puts a hard limit of 0.1 on the eddy viscosity in the lowest 200 meters of the CRM.

Also cleaned old up experimental stuff. Removed two version of SP_FLUX_MOD and SPFLUXBYPASS (version 3 and 4) since these have not shown any success in improving the SP simulations. The code for version 1 and 2 of SPFLUXBYPASS were left in place, but everything was commented out in case we want to revisit this issue in the future.

Changed the timescale of the surface flux smoother to 4 hours instead of 2, and removed my previous approach with an adaptive timescale.