Misc bug fixes
In addition to some bugs and typos that were causing compiler erros, I also added an SP_PHYS_BYPASS option to bypass some problematic physics routines that are run alongside the CRM, but the tendencies are not used. I set this to get around a segfault being caused by the MG microphysics scheme, but it seems that there might be a floating point exception caused elsewhere. Still need to find a way to trace this down.