increased depth of minimum TK layer
The eddy viscosity (TK) limiterin the CRM appears to help the near-surface strato-fog issue that occurs in SP simulations. The layer where the minimum TK is enforced was increased to 1 km. The value is still hard coded as 0.1. If further testing shows promising results, the depth and value of this limiter will be set as namelist variables, but for now we will leave this hardcoded to be activated with the preprocessor variable SP_TK_LIM.