Add infrastructure to read coefficients from file
Add new class and method to read the gas optics coefficients from the
provided RRTMGP netCDF files. The class holds the fields from the file,
as well as provides a type-bound method for reading the file, allocating
space to hold the data, and then carrying this data throughout the
module to be accessed where it is needed before the calls to the
radiative transfer methods. This is a replacement for the corresponding
subroutine provided with RRTMGP that both reads the data and initializes
a ty_gas_optics_specificiation object. We replace this because
initializing the ty_gas_optics_specification object now also requires
passing a ty_gas_concs object to the routine. This is a problem for our
implementation because we need access to the physics_state and
physics_buffer_desc objects to initialize the gas concentrations in
radiation_init, but these are not passed to the routine, and doing so is
somewhat complicated by the fact that in physics_init, the
physics_buffer_desc pointer is two-dimensional as it contains all of the
"chunks". Initializing the gas concentrations also involves calls to
rad_cnst_get_gas, which takes an argument specifying the value of icall
over N_DIAG "climate/diagnostic calls". I think it would be reasonable
to pass icall=0 as a dummy argument, but this is done later in
radiation_tend over all icall from 0 to N_DIAG when loading gas
concentrations for each pass. So instead, we just load the data from the
netCDF file into memory and keep it there until we need it.

This commit also fixes some missing configuration in the build for
RRTMGP. The model now seems to at least build with the rad=rrtmgp
option, although radiation_tend has still yet to implement most of the
actual functionality (i.e., running the model will not doing any
radiative transfer yet). Still to be done include writing routines to
load gas, cloud, and aerosol optical properties, perform the McICA
sampling, and translate data between RRTMGP objects and CAM state, pbuf,
and output fields.