Altered Eddy Viscosity Limiter
After some discussion I decided to remove the positive definite-ness that was previously imposed on the buoyant production term used for the eddy viscosity. This was our initial attempt to promote more turbulence in the boundary layer, but now that we are using a hard limiter it seems pointless to enforce this term to be positive definite.

Another point of discussion was that using a limit of 0.1 over a depth of 1 km for the hard limiter seemed to be too much. Further test runs with a limit of 0.05 over 500m showed a similar impact on the near-surface cloud problem and somewhat less of an impact on the lower tropospheric cloud fraction. So I've left the model in this configuration. The value and depth of the limiter are set using the variables tk_min_value and tk_min_depth, which are hard coded at the moment, but can be easily altered to be set by namelist variables in the future if needed.