Stared working on ESMT PGF
This is the start of work to get a working code to calculate the pressure gradient force for the scalar momentum transport. This is the trickiest part, so it will involve several interim commits. The cpp variable ESMT_HIDE_PGF is used to hide the pgf code so I don't have to leave it commented out and can still run the model withou the PGF part.