Added Workarounds for specifi SP issues
Two issues came up in long SP runs, which are remedied by this commit.

The first was that the log files were unreasonably large due to lots of QNEG errors in the layers above the CRM. The files were so big that I was hitting the disk space quota quicker than expected. To reduce the size of these log files, I added a condition in chemistry/aerosol/wetdep.F90 that sets a tracer value to zero if a small negative number is found above the CRM top, only for SP runs. This should not affect standard E3SM runs.

The second issue was that the model would occassionally fail due to the UW moist turbulence scheme trying to create cloud layers in the upper stratosphere, which would then be extrended past the model top by zisocl(). Since this is no realiztic, I decided to limit the base of these cloud layers to only occur below 50 mb. This was done in the exacol() function of eddy_diff.F90, which I believe stands for "examine the column" and identify cloud layers. The fix seems to have worked, and the run that was failing is now continuing without issue. Since the standard E3SM does not use the UW moist turbulence scheme, this will nto affect the solution. I also do not expect it to affect the solution of the SP model in any significant way, since this scheme is only needed indirectly for aerosol transport and processing.