Working version of ESMT w/o PGF
A lot of modifications were needed to get the basic ESMT working. There was a problem in how I originally defined the dimensions of the scalar momentum variables. So the current version advects and diffuses the scalar momentum variables, but the pressure effects are still a work in progress.

I also was having trouble getting resolved momentum tendencies out of some 3D CRM runs, and it turned out that the variable CRM3D was not defined by default, but YES3D is. After examining the code it was clear that the only use of CRM3D was to remove the momentum tendencies from the CRM interface, even tho the tendency values were not returned without definiing SPMOMTRANS. SO I decided to remove all "ifdef CRM3D" checks to simplify the code. If passing these two variables in the CRM interface is associated with a performance penalty, then we can add these ifdefs back in.

In the process of debuggin I ended up removing the ifdef that bypassed the GCM diffusion of momentum, so there is some double counting of momentum currently. I'll add this back in a future commit.