Added experimental option for CRM Re-Initialization
The new option re-initializes CRM fields to explore whether this is a viable strategy for reducing communication overhead when we move to running the CRM on the GPU. The CPP variables for enabling this feature are listed below. Note that the SP_CRM_REINIT flag is needed but does not enable any re-initialization on it's own, so it must be combined with one of the other variables. This was the cleanest way I coudl think of for testing different variable combinations without making lots of nearly identical code blocks.

  SP_CRM_REINIT     - Turn on CRM reinitialization
  SP_CRM_REINIT_UV  - re-initialize horizontal velocity
  SP_CRM_REINIT_W   - re-initialize vertical velocity
  SP_CRM_REINIT_T   - re-initialize temperature
  SP_CRM_REINIT_Q   - re-initialize moisture variables (including condensate)