Streamline cospsimulator_intr_run
Streamline cospsimulator_intr_run in preparation for adding code to
handle CRM columns. The three duplicate blocks that setup and call the
simulators for different cases have been reduced to one. This was used
before to allow separate calls for running all simulators on all
columns, running passive simulators only on daytime columns, and
allowing for A-Train sampling for active simulators. We now have just a
single block to setup and run the simulators. Currently, we do not use
the A-Train sampling functionality (and I do not think it was ever
rigorously tested or even used), but this should be easily added column
by column when we add the loop over columns to handle the CRM case. The
new code now also runs the simulators on both day and night columns, but
we not properly set the sunlit flag, which will correctly mask the
nighttime columns for those simulators that require it (ISCCP, MISR, and
MODIS). It would also be straightforward when looping over ncol to turn
off these simulators for individual columns based on the value of
sunlit, which might provide a very slight performance gain, but the
passive simulators are not the computationally expensive part of the
code to begin with.