Fix timestep for SP runs at ne4 resolution
Fix timestep for SP runs at ne4 resolution so that SP tests will run to
completion within wallclock limit on short queue on titan. The default
ne4 configuration is to run with a long timestep, but this seems to make
for unstable runs when configured with SP on. At least a temporary fix
seems to be to turn down the timestep, which is done by increasing the
value of ATM_NCPL. We do this just for cases configured with SP on
(using the SP compsets) and at a%ne4 resolution (the a%ne4 non-SP cases
should be unaffected).