Fixes for SP_FLUX_BYPASS
My previous implementation of the SP_FLUX_BYPASS option had applied the fluxes outside of crm_physisc_tend(), but I had forgotten that the state is reset within this subroutine, so the surface fluxes were effectively zero. This commit fixes this problem and removes various other code I used for testing. I also added a second call to qneg4() in tphysbc, which adjusts the fluxes in cases of strong negative latent heat flux. The second call is needed since the environment has been modified by the dynamics. I have left the first call in place without bypassing it.

There is also only 1 version of the flux bypass now. At first I implemented a second version that would also bypass all constituent fluxes. This option is gone now, but we might have to revisit this when we get ECPP working.