Add code to run lw gas optics
Continued development of the RRTMGP implementation allowing for running
the longwave driver now with the gas optics loaded. Cloud and aerosol
optics are still zeroed out, and the gas optics seem to be failing for
some reason, but this commit has the added logic and initialization to
get this running hopefully after a few bug fixes. The gas coefficient
initialization and read routines in the previous commit have also been
replaced with a modified version of the RRTMGP load_and_init routine
from mo_load_coefficients module, which is provided with RRTMGP as an
example for users.

NOTE: this version of the code will not build with -rad rrtmg at the
moment because the radiation initialization has been moved to a separate
subroutine within the radiation module, rather than being handled by the
physics initialization routine. This is because RRTMGP requires a few
different options from the older packages. This needs to be made
consistent with the other radiation packages somehow, either by changing
the options to be consistent, or by writing separate initialization
routines for RRTMG and CAMRT.