Fixed Eddy Diffusivity issue
For the case with SP_FLUX_BYPASS enabled I was running into a run-time crash because the fluxes are needed for the caclulation of eddy diffusivity. This calculation calls the routine compute_vdiff() which is where the fluxes are applied (although I'm not sure why). So I needed to add an option to bypass the SP_FLUX_BYPASS for this specific call.