Fixing but in micro_init
So, there are diffs for some of the micro_fields, and that concerned me. When I run -O0 and -O3 with GNU, I don't see these diffs. However, when I run -O3 with GNU and -O3 with PGI, I do see diffs in the exact same fields (and of much higher magnitude than I see between the redimensioned code and the original code). Also, I tried exiting the main time step loop after one step to compare redimensioned against master, and after one time step, the diffs are only a max of 10^-7. This together tells me the port is probably fine and that we're almost certainly just seeing machine precision changes amplifying over time. I suppose this inherently means the M2005 microphysics are significantly more non-linear than the SAM1MOM microphysics. But M2005 is now in the redimensioned code.