Fix bug setting col_dry in gas optics
Fix bug in gas optics in section of code dealing with column gas amounts
when col_dry is not passed as an input. The problem was that this
section of code was trying to use a string array that was assumed to be
allocated and assigned in a block of code above, but that section of
code was effectively commented out by an if (.false.). This was causing
a string search on a garbage array, result in a negative index that
filled the col_dry array with NaNs. This commit fixes this issue, and
allows RRTMGP to run with CAM inputs on gas optics. Still TODO is turing
on aerosol optics and implementing the cloud optics calculations.