Fix character length bug in RRMTGP optics init
Fix bug in RRTMGP optics initialization routines where the optionally
input "name" strings were mistakenly assumed to be single characters
rather than strings (character arrays). This meant that the "name" field
was populated with only the first character of the input string. This
resulted in less than helpful error messages when something went wrong
with an optics object that then could not completely report itself with
the name label.