Fixed SP test issue with COMPARE_base_rest
The ERP tests of the SP compsets were failing at the COMPARE_base_rest stage. The differences were widespread and it was difficult to track down what was causing them. In an unrelated effort for implementing a random walk to make the CRM orientation evolve over time I stumbled onto a concerning discrepancy between the behavior of the fortran intrinsic random number generator (RNG) when compiling with gfortran and pgi. The results with the pgi compiler showed that the random numbers were scaled by the seed value, and so small seed values resulted in near zero random values. The solution is to build in a RNG so that the results are consistent across compilers. Since I was already doing this, I figured I'd see whether the test results would be affected by using this new RNG to set the CRM perturbations in setperturb(). I also changed the condition that decides whether to call setperturb() and based it on the logical "is_first_step" instead of the old method that compared a few arbitrary CRM cells to see if they were equal. These changes resulted in a passing test of COMPARE_base_rest, however I'm lacking a complete explanation of why this worked. In theory, neither change chould have affected this comparison as long as the intrinsic RNG got the same seed and gave the same results, even if those values were near zero. The difference may lie in how the intrinsic RNG was called such that the string of random numbers was "remembered" from the previous seed, in which case changing the processor count may result in a different sequence. Either way, the new RNG is much more robust and reliable. The new RNG uses the Mersenne Twister algorithm. There is already a Mersenne Twister RNG implemented elsewhere in the model, but it is cumbersome to get single values, so I think the current approach is preferable for the CRM.