Added SP_CRM_CPLIT and SP_USE_DIFF options
This commit is the result of more tinkering and analysis of the CRM coupling strategy that started with the SP_FLUX_BYPASS option and expanded into other things (which should have been broken into other branches - my bad). Some new options were added, some were the result of recent discussions at a GASS conference. I've outlined the new options below, which are mostly preprocessor variables to ensure the default model behavior is preserved.

SUPPRESS_QNEG_MSG - this just stops the QNEG3 warnings from showing up in the log file because they were making it hard to debug certain failures.

SP_USE_DIFF - This activates the GCM's standard vertical diffusion routines. I initially thought this would be "double counting" the diffusion, but Marat convinced me that it's better to think of this as "co-working" between the GCM and CRM diffusion. This option does seem to stabilize the model a little bit. However, it doesn't affect the climate very much, which you might view as good or bad.

SP_CRM_SPLIT - This option splits the CRM integration in half. One half is done in the normal place after the dynamics (i.e. tphysbc), and the other half is run at the end of tphysac so that surface fluxes can be diffused by the CRM before the dynamics is called. I've had mixed results testing this. The two calls to the CRM seems to fight back and with the tendencies near the surface, but this may be a subtle error in how I've set this up. I want to leave this in for now and keep exploring this option.

A couple other miscellaneous changes are included since they helped suppress various failure modes.