Changed dt_crm to crm_run_time for clarity
dt_crm was implemented along with the SP_CRM_SPLIT option to represent the total integration time of the CRM, which was not necessarily equal to the GCM timestep. Using "dt" in the variable named seemed potentially confusing, so I opted for a more descriptive name of crm_run_time, and icrm_run_time for its inverted value.