Working version of ESMT PGF and SP_ORIENT_RAND
This commit includes some major changes to the explicit scalar momentum transport (ESMT) code that enabled calculation of the cloud scale pressure gradient force (PGF). The PGF needed for the scalar momentum traces is different than the PGF normally calculated for the momentum on the CRM grid. I decided to add an FFT library to be consistent with Stefan Tulich's implementation of this idea in SP-WRF. However, the ideal implementation of doing the FFT calculation on all columns at once did not work. The reason was a confusing error from deep in the FFT library about reading outside of the array bounds. I spent several weeks trying to understand this, but in the end I couldn't explain it, so I opted to just calculate the FFT over each level separately, which worked fine on the first try.

The other part of this commit implements the SP_ORIENT_RAND option, which was another idea from Stefan Tulich at NOAA. The main idea is to have each 2D CRM change orientation using a random walk in angular space in order to avoid biasing the climate to any particular CRM orientation. Preliminary results show that this can smooth out the erratic precipitation pattern (but only slightly). This only affects the winds that are used as input to the CRM. The random walk is configured such that the CRM will rotate 90 degrees per day on average. This is done with the crm_rotation_offset parameter, which essentially shifts the distribution of angular steps away from zero. The basic idea is to keep the CRM rotating instead of twitching back and forth around the same initial orientation angle. Even without the SP_ORIENT_RAND option, the CRM orientation is specific with an angle instead of directly using the GCM U and V wind, which dramatically simplified how the SP_DIR_NS option works. This new framework can also be use for 3D CRM configurations such as 32x4.