Add LW cloud optics and aerosol
Add code to handle LW cloud optics and aerosol optics, and have RRTMGP
calculate fluxes from combined optical properties. This involved a few
bug fixes internal to RRTMGP. In particular, the optics%increment()
functions seemed to fail when passing aerosol optical properties by
band...the increment function was trying to infer the bands from the
wrong input (the one that should have contained gpoints, not bands). I
also added some debugging checks into a lot of the RRTMGP source to try
to figure out what was going on. This should be removed in a future
commit (or update the RRTMGP source to the new version).