Fixes for build errors while merging PR #28
When merging PR #28 the regression tests on Edison with the intel compiler revealed some build errors that didn't show up in the initial testing with PGI on Titan. The phys_stage variable needed to be defined in tphysbc() and tphysac() because the phys_stage argument for crm_physics needs to have intent(inout), so a variable needed to be passed instead of a value. There was also an error of variablesn crm_module.F90 being defined twice that was inadvertantly created by the merge.

 Changes to be committed:
	modified:   cime/config/e3sm/machines/config_machines.xml
	modified:   components/cam/src/physics/cam/physpkg.F90
	modified:   components/cam/src/physics/crm/crm_module.F90