Moved initialization of coupler precip variables
A test run of the SP_PHYS_BYPASS revealed NaN's in the ice model that originate in crm_physics.F90 due to the code structure dictated by the "is_first_step" flag. The pointers used for the output precip variables used by the coupler, such as prec_dp, are only initialized when is_first_step is false, so the model will crash on the first step if the off-line physics, like ZM and UW convection schemes, are bypassed. I've always been suspect of how this part of the cod eis organized, so there may be other problems like this.

I also modified the indentation and added comments to many of the #endif statements for clarity.

	modified:   components/cam/src/physics/cam/crm_physics.F90