Update RRTMGP to latest version
Update RRTMGP to latest version on the RRTMGP trunk. This actually
required a few bug fixes to get to work. First, the load_coefficients
routine needed to be replaced with my old modified version, which
explicitly allocates space for the fields in the kdist objects.
Otherwise, these seem to fall out of scope for some reason and get
filled with garbage before being used. Also, some changes were needed in
the optics "increment_by" routines. These seemed to be treating inputs
improperly for cases where optical properties were passed by band,
rather than by g-point. I submitted the bug to Robert Pincus to
hopefully be addressed in the trunk, so we can pull in that change when
it's applied.