Fixes for rebase on master
Recent changes on master changed the requirements on the radiation
interface, which were not taken into account when the RRTMGP interface
was being developed. Specifically, the aerosol radiative properties
routines now require knowing what kind of file is used to prescribe
volcanic aerosol. The way this was added in the chemistry interface ad
in the aer_rad_props interface made this kind of awkward, requiring
passing a flag through a number of levels of subroutines. So, I changed
this behavior slightly to instead make the flag a module variable in the
chemistry interface so that we could just look it up from aer_rad_props,
rather than passing it down from subroutine to subroutine.